[ci-visibility] Allow using `after:run` directly  (#4090)
