+++ ci/cypress/after-run.js
@@ -1,0 +1,1 @@
+module.exports = require('../../packages/datadog-plugin-cypress/src/after-run')
