+++ packages/datadog-plugin-cypress/src/after-run.js
@@ -1,0 +1,3 @@
+const cypressPlugin = require('./cypress-plugin')
+
+module.exports = cypressPlugin.afterRun.bind(cypressPlugin)
